capture filesystem access needed to create a nested sandbox
Make

 #lang racket/base
 (require racket/sandbox)

 (define e (make-module-evaluator (string-append "#lang racket/base\n")))
 (e '(require racket/sandbox))
 (e '(make-module-evaluator (string-append "#lang racket/base\n")))

work for a Git-checkout build, where parsing the collection links file
requires resolving relative paths that contain up-directory indicators.

The test for this repair is in the "racket-test-core" collection.
(Arguably, that test should be moved, but the sandbox test suite
turns out to be an especially valuable test for core functionality.)

Relevant to racket/racket#1549