yet more repairs to the interaction of errors and `let-values`
Continuing the saga that includes 8190a7730d and d1ba9fbb6e, it turns
out that a 0-binding clause as the last one isn't so special after
all. A little later in the optimizer, now that we're sometimes moving
an error to the body, we can't assume that the body can be discard
if an error was detected.

(cherry picked from commit 736cdfb2c18a91da64498d04944a9ae1e5fcdd82)