repair for `hash-keys-subset?`
When comparing a part of a hamt that is a collision node versus a
subtree node, a "hash code" was extracted from the collision node ---
but that's really a code for an integer key is that used for the
collision element. The comparison should instead use a code extracted
from the reference to the collision node (which is the code that is
common to all colliding keys).

(cherry picked from commit 6de3659664be532bd195ca9b1abcb2b825824de2)