changed make-variable-like-transformer to create #%app form
The docs claim that "Uses of the macro in operator position are
interpreted as an application", yet the current implementation
confusingly wraps #%expression around the operator instead of just
creating #%app.

Worse, the #%expression isn't in the phase level of the application,
meaning that languages can't easily use make-variable-like-transformer
if they intend to reimplement #%app. Maybe I'm mistaken and there is
rationale behind doing it this way / code that depends on the current
behavior. I find my self needing this macro but just rewriting it this
way since the bad phase-level #%expression just gets in the way and I
wanted #%app instead.