syntax/parse: remove some absolute paths related to lazy-require-syntax
Since the previous version of this code was first written, Racket's
requirements seem to have relaxed---but not completely!

If *all* related module paths are made relative, then the deps
test case breaks (but only the test is pre-compiled ?!!).

If requires from disappearing module(s) to residual module(s)
use absolute module paths, it seems to work.