Small update to move build out to a seperate file, also adding expected summary.
Moved command out of travis configuration file and created a separate bash
script we could run.  The filtering appears to be getting in the way of the
error code being forwarded up to the running shell, and it is possible the -e
option on bash will help.  The summary file is the expected result file (for
diffing, if necessary).
  .travis.yml, .travis/dobuil.sh (new), .travis/summary

original commit: 48cb144d83b5fffa0d2523790a6d0bf172d154ea